moved the frepytools time-average generation code into the class implementation
